requirements.txt
aniso8601==8.0.0
attrs==19.3.0
blinker==1.4
boto3==1.10.22
botocore==1.13.22
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-Script==2.0.6
gunicorn==20.0.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==7.2.0
pkg-resources==0.0.0
pymongo==3.9.0
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
